Consultas de Medicos

+++ app/Http/Controllers/Admin/DoctorController.php
@@ -7,15 +7,22 @@ use App\User;
 
 use App\Http\Controllers\Controller;
 use App\Specialty;
+use App\WorkDay;
+use Carbon\Carbon;
+use Illuminate\Support\Facades\Validator;
 
 class DoctorController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    
+    private $days = [
+        'Lunes',
+        'Martes',
+        'Miércoles',
+        'Jueves',
+        'Viernes',
+        'Sábado',
+        'Domingo'
+    ];
+
     public function index()
     {
         $doctors = User::doctors()->paginate(10);
@@ -42,29 +49,52 @@ class DoctorController extends Controller
      */
     public function store(Request $request)
     {
+
         $rules = [
-            'name'      =>  'required | min:3',
-            'email'     =>  'required | email',
-            'ci'        =>  'nullable | digits:6',
-            'address'   =>  'nullable | min:5',
-            'phone'     =>  'nullable | min:6'
+            'name'      => 'required|min:3',
+            'email'     => 'required|email|unique:users',
+            'ci'        => 'required|unique:users,ci',
+            'address'   => 'nullable|min:5',
+            'phone'     => 'nullable|min:6'
         ];
-        $this->validate($request, $rules);
+
+        $messages = [
+            'required'  => 'El campo :attribute es requerido.',
+            'unique'    => 'Ya existe el :attribute en nuestra base de datos.',
+            'email'     => 'El :attribute debe ser una dirección de correo electrónico válida.',
+            'min'       => 'El campo :attribute debe tener al menos :min caracteres.'
+        ];
+
+        $attributes = [
+            'name'      => 'Nombre',
+            'email'     => 'Correo electrónico',
+            'address'   => 'Dirección',
+            'phone'     => 'Teléfono',
+            'ci'        => 'Cédula de Identidad'
+        ];
+
+        $validator = Validator::make($request->all(), $rules, $messages, $attributes);
+
+        // Verificar si la validación falla
+        if ($validator->fails()) {
+            return redirect()->back()->withErrors($validator)->withInput();
+        }
 
         $user = User::create(
             $request->only('name', 'email', 'ci', 'address', 'phone')
-        + [
-            'role'      =>  'doctor',
-            'password'  =>  bcrypt($request->input('password')),
-        ]);
+                + [
+                    'role'      =>  'doctor',
+                    'password'  =>  bcrypt($request->input('password')),
+                ]
+        );
 
         $user->specialties()->attach($request->input('specialties'));
-        $notification = 'El médico ha sido registrado correctamente';    
+        $notification = 'El médico ha sido registrado correctamente';
 
 
         return redirect('doctors')->with(compact('notification'));
     }
-        
+
     /**
      * Display the specified resource.
      *
@@ -74,6 +104,29 @@ class DoctorController extends Controller
     public function show($id)
     {
         //
+    }
+
+    public function showHours(Request $request, $id) {
+        $workDays = WorkDay::where('user_id', $id)->get();
+        if(count($workDays) > 0) {
+            $workDays->map(function ($workDay) {
+                $workDay->morning_start = (new Carbon($workDay->morning_start))->format('g:i A');
+                $workDay->morning_end = (new Carbon($workDay->morning_end))->format('g:i A');
+                $workDay->afternoon_start = (new Carbon($workDay->afternoon_start))->format('g:i A');
+                $workDay->afternoon_end = (new Carbon($workDay->afternoon_end))->format('g:i A');
+                return $workDay;
+            });
+
+        }else{
+            $workDays = collect();
+            for ($i = 0; $i<7; ++$i)
+            $workDays->push(new WorkDay());
+        }
+        
+        $days = $this->days;
+
+         
+        return view('schedule', get_defined_vars());
     }
 
     /**
@@ -88,7 +141,7 @@ class DoctorController extends Controller
         $specialties = Specialty::all();
 
         $specialty_ids = $doctor->specialties()->pluck('specialties.id');
-        return view ('doctors.edit', get_defined_vars());
+        return view('doctors.edit', get_defined_vars());
     }
 
     /**
@@ -101,28 +154,38 @@ class DoctorController extends Controller
     public function update(Request $request, $id)
     {
         $rules = [
-            'name'      =>  'required | min:3',
-            'email'     =>  'required | email',
-            'ci'        =>  'nullable | digits:7',
-            'address'   =>  'nullable | min:5',
-            'phone'     =>  'nullable | min:6'
+            'name'      => 'required|min:3',
+            'email'     => 'required|email',
+            'ci'        => 'nullable|digits:7|unique:users,ci,' . $id,
+            'address'   => 'nullable|min:5',
+            'phone'     => 'nullable|min:6'
         ];
+
+        // Validar los datos recibidos del formulario
         $this->validate($request, $rules);
+
+        // Buscar al médico por su ID
         $user = User::doctors()->findOrFail($id);
+
+        // Obtener los datos del formulario
         $data = $request->only('name', 'email', 'ci', 'address', 'phone');
+
+        // Verificar si se proporcionó una nueva contraseña
         $password = $request->input('password');
-        
-        if ($password)
+        if ($password) {
             $data['password'] = bcrypt($password);
+        }
+
+        // Actualizar los datos del médico
         $user->fill($data);
         $user->save();
 
+        // Sincronizar las especialidades del médico
         $user->specialties()->sync($request->input('specialties'));
 
-        $notification = 'La información del médico se ha actualizado correctamente';    
-
+        // Redireccionar con una notificación
+        $notification = 'La información del médico se ha actualizado correctamente';
         return redirect('doctors')->with(compact('notification'));
-        
     }
 
     /**
@@ -134,9 +197,18 @@ class DoctorController extends Controller
     public function destroy(User $doctor)
     {
         $doctorName = $doctor->name;
-        $doctor->delete();
-        $notification = "Doctor $doctorName eliminado de la BD correctamente";
 
+        // Verificar si el doctor tiene días agendados
+        if ($doctor->workDays()->exists()) {
+            $notification = "No se puede eliminar al Doctor $doctorName porque tiene días agendados previamente.";
+
+            return redirect(route('doctors.index'))->with('danger', compact('notification'));
+        }
+
+        // Si no tiene días agendados, proceder con la eliminación
+        $doctor->delete();
+
+        $notification = "Doctor $doctorName eliminado de la BD correctamente";
         return redirect(route('doctors.index'))->with(compact('notification'));
     }
 }
